moves doscstrings to folder
