+++ feature_engine/creation/cyclical.py
@@ -5,7 +5,7 @@ import pandas as pd
 from sklearn.utils import deprecated
 
 from feature_engine.base_transformers import BaseNumericalTransformer
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
