+++ feature_engine/creation/math_features.py
@@ -4,7 +4,7 @@ import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.creation.base_creation import BaseCreation
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
