+++ feature_engine/discretisation/arbitrary.py
@@ -7,7 +7,7 @@ from typing import Dict, List, Optional, Union
 import pandas as pd
 
 from feature_engine.discretisation.base_discretiser import BaseDiscretiser
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
