+++ feature_engine/encoding/ordinal.py
@@ -5,7 +5,7 @@ from typing import List, Optional, Union
 
 import pandas as pd
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
