+++ feature_engine/encoding/woe.py
@@ -6,7 +6,7 @@ from typing import List, Union
 import numpy as np
 import pandas as pd
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
