+++ feature_engine/imputation/categorical.py
@@ -6,7 +6,7 @@ from typing import List, Optional, Union
 import pandas as pd
 
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
