+++ feature_engine/imputation/missing_indicator.py
@@ -7,7 +7,7 @@ import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
