+++ feature_engine/outliers/artbitrary.py
@@ -11,7 +11,7 @@ from feature_engine.dataframe_checks import (
     _check_contains_na,
     check_X,
 )
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_not_learn_docstring,
