+++ feature_engine/outliers/trimmer.py
@@ -4,7 +4,7 @@
 import numpy as np
 import pandas as pd
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
