+++ feature_engine/outliers/winsorizer.py
@@ -7,7 +7,7 @@ import numpy as np
 import pandas as pd
 
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
