+++ feature_engine/selection/drop_duplicate_features.py
@@ -3,7 +3,7 @@ from typing import List, Union
 import pandas as pd
 
 from feature_engine.dataframe_checks import _check_contains_na, check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
