+++ feature_engine/selection/recursive_feature_elimination.py
@@ -1,7 +1,7 @@
 import pandas as pd
 from sklearn.model_selection import cross_validate
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
