+++ feature_engine/selection/single_feature_performance.py
@@ -5,7 +5,7 @@ import pandas as pd
 from sklearn.model_selection import cross_validate
 
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
