+++ feature_engine/selection/target_mean_selection.py
@@ -7,7 +7,7 @@ from sklearn.model_selection import cross_validate
 from feature_engine._prediction.target_mean_classifier import TargetMeanClassifier
 from feature_engine._prediction.target_mean_regressor import TargetMeanRegressor
 from feature_engine.dataframe_checks import check_X
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _feature_names_in_docstring,
     _fit_transform_docstring,
