+++ feature_engine/timeseries/forecasting/expanding_window_features.py
@@ -6,7 +6,7 @@ from __future__ import annotations
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
