+++ feature_engine/timeseries/forecasting/window_features.py
@@ -3,7 +3,7 @@ from typing import Callable, List, Optional, Union
 import pandas as pd
 from sklearn.utils.validation import check_is_fitted
 
-from feature_engine.docstrings import (
+from feature_engine._docstrings.docstrings import (
     Substitution,
     _drop_original_docstring,
     _feature_names_in_docstring,
